Optimized BFS
My redPac AI(whis is suppose to be trying to eat the Ghost) was not very smart. It would run into blocks and stay there. The way I implemented the AI at first was by getting the difference in coordinates from my source to the Ghost and using that difference to make a decision. I changed it to BFS. To do this I needed to create a Node(Cell) representation of my 2D matrix. Then find the BFS, by keeping track of distance for each cell and have every Cell have a Node to a parent Cell to be able to backtrack my path.