2 new methods
Added a removeObject() to remove objects based on input. Also, when a pacMan moves, to check if it is in the same coordinate as Ghost, i need to lookup the Ghost object. 
The changeIds() method is called when my regular pac-men are eaten. Its purpose is to change the Id of the redPacs so that its AI changes to that of a regular pac.